Improve Security Logging
Whenever a request is made, the request has to undergo security
filtering to verify that the client is permitted to access the resource
they wish to receive. When a user attempts to access a request without
supplying credentials, their request is rerouted to the
AuthenticationEntryPoint, where it is intercepted and a 401 status code
is returned.

Whenever commence() is called in AuthenticationEntryPoint, a lengthy
stack trace is printed to the console, unecessarily. This clutters the
logs and make it difficult to see more important things.

Instead, log at the WARN level instead of ERROR, and print a more useful
message, without a stacktrace.