Integration with AWS S3 for Picture Storage
Users can now upload profile pictures using the new API for user
image upload. Images uploaded to the server are uploaded to AWS S3 using
the AWS JDK for Java.

The client can upload profile pictures by making a multipart request to
the URI '/users/{id}/profile/imageupload?file=filename.jpg'.

To use the new API, developers must obtain AWS IAM user credentials and
create environment variables to supply those credentials to the
AmazonS3ClientService. This is documented in the README.md.

The AWS SDK for Java was added as a dependency for the project. A new
@Service was created 'AmazonS3ClientService' which is used to
communicate with AWS. The service was designed to be extensible so that
it can be used for other purposes in the future.

When a user uploads a profile picture and the image is uploaded to S3,
the object key is stored in the User's model. When the user is fetched
through UserService#adaptUserToSummmary(User user), a presigned URL is
generated and attacked to the UserSummaryResponse.

Some new configuration properties were introduced. These changes are
documented in the DeploymentInfrastructure.md document.