Refactor of Service and Controller Layer
Major refactor of the service layer and controllers to reduce coupling
between them. Now, it is easier for service layer entities to reuse
logic.

These changes allowed me to remove DAO dependencies on in various
service layer entities, and instead use the appropriate service layer
entity, which helped to improve service layer cohesion.

I also fixed a bug that caused deleting users, buckets, items, and
relationships to fail due to a database constraint. In the process, I
also fixed a bug that caused private buckets to appear in queries due to
an oversight in the custom DAO queries for the user feeds.