Enabled Actuator Endpoints
Enabled Spring actuator endpoints, and adapted the deployment scripts to
make use of the endpoints to graceful stop the server and validate the
server once started.

Also created a database initialization script for use in production,
which ensures that an administrator user exists in the database so that
the code deploy agent can authenticate during the deployment.

Created another authentication endpoint /auth/authenticate, which is a
more script-friendly authentication mechanism.

Fixed user roles so that the hasRole("ADMIN") would work correctly.